Add code for slow movement
Added code to the #move method that puts a 5 second delay between the 
elevator moving from floor to floor, to simulate the actual motion of a 
real life elevator.

Also added code to check whether the elevator was going up or down.  
This was needed so that the program could correctly calculate the floors 
to be traversed and output the correct time intervals.